+++ src/tallybin.h
@@ -1,5 +1,5 @@
 #ifndef __TALLYBIN_H__
 #define __TALLYBIN_H__
 #include "material_data.h"
-unsigned readbins(double** binedges, char* input);
+unsigned readbins(float** binedges, char* input);
 #endif
